working -> and now the bonus taks

+++ 08.InterfaceSegregationDependencyInversion/InterfaceSegregationDependencyInversionHomework/src/app/contracts/Boat.java
@@ -4,7 +4,7 @@ public interface Boat extends Modelable {
 
     int getWeight();
 
-    int getSped();
+    double getSped();
 
     boolean hasMotor();
