+++ 08.InterfaceSegregationDependencyInversion/InterfaceSegregationDependencyInversionHomework/src/app/contracts/Engine.java
@@ -2,6 +2,6 @@ package app.contracts;
 
 public interface Engine extends Modelable {
 
-    int getOutput();
+    double getOutput();
 
 }
